+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - [QA] Simplify Playground endpoint
 - [QA] Reorganize e2e tests around scopes
+- [REFACTOR] Introduce common LiveBlock component
 
 ## 0.10.1